+++ lib/sg_tiny_backup/commands/base.rb
@@ -15,7 +15,7 @@ module SgTinyBackup
         [0]
       end
 
-      def strong_warning_message
+      def warning_message
         nil
       end
     end